+++ src/peach_discord_client/Dockerfile
@@ -1,6 +1,7 @@
 FROM golang:1.13.8 as builder
+RUN go get -t github.com/stretchr/testify
 COPY . .
-RUN go get -d -v ./...
+RUN go get -d ./...
 RUN CGO_ENABLED=0 go build -o /app -a -ldflags '-extldflags "-static"' .
 
 FROM alpine:3.11
