+++ src/peach_shard_coordinator/main.go
@@ -1,55 +1,33 @@
 package main
 
 import (
-	"encoding/json"
-	"fmt"
-	"log"
 	"net/http"
 	"os"
 
 	"github.com/gorilla/mux"
+
+	log "github.com/sirupsen/logrus"
 )
 
 var shards []shard
 var bottoken string = os.Getenv("BOTTOKEN")
 
-// function used to create the shards object and to fetch the shard amount
-func resetShardCount() {
-
-	// fetch recommended shard amount from discord api
-	client := &http.Client{}
-	var response gatewayresponse
-	req, err := http.NewRequest("GET", "https://discordapp.com/api/gateway/bot", nil)
-	if err != nil {
-		log.Fatal(err)
-	}
-	req.Header.Add("Authorization", fmt.Sprintf("Bot %v", bottoken))
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer resp.Body.Close()
-	json.NewDecoder(resp.Body).Decode(&response)
-
-	// create list with shard objects
-	shardCount := response.Shards + 1
-	shards = make([]shard, (shardCount))
-
-	// set shardIDs and roles
-	DMshard := true
-	for shardID := 0; shardID < shardCount; shardID++ {
-		if DMshard {
-			shards[shardID].ShardID = shardID
-			DMshard = false
-		} else {
-			shards[shardID].ShardID = shardID - 1
-			shards[shardID].Server = true
-		}
-	}
+func init() {
+	// Set log format, output and level
+	log.SetFormatter(&log.TextFormatter{
+		ForceColors:      true,
+		PadLevelText:     true,
+		QuoteEmptyFields: true,
+		DisableTimestamp: false,
+		FullTimestamp:    true,
+		TimestampFormat:  "2006-01-02 15:04:05",
+	})
+	log.SetOutput(os.Stdout)
+	log.SetLevel(log.InfoLevel)
 }
 
 func main() {
-	log.Printf("shard coordinator online\n")
+	log.Info("shard coordinator starting...")
 
 	// setup paths
 	r := mux.NewRouter()
@@ -59,9 +37,16 @@ func main() {
 	api.HandleFunc("/updateshard", updateShard).Methods(http.MethodPost)
 	api.HandleFunc("/scale", scale).Methods(http.MethodGet)
 
-	// initial creation
+	// initial creation of shards list
 	go resetShardCount()
 
 	// run
-	log.Fatal(http.ListenAndServe(":8080", r))
+	done := make(chan bool)
+	go http.ListenAndServe(":8080", r)
+
+	// log ready
+	log.Info("shard coordinator online")
+
+	// wait for goroutine to finish
+	<-done
 }
